refactor(tests): move public-API collection tests from inline mod to tests/cli_collection_*.rs
Migrates 26 tests out of `src/commands/collection.rs::mod tests` into 6
per-feature integration test files under `tests/`, leaving 59 white-box
tests inline with a module-level annotation listing the private items they
exercise. Pure mechanical refactor: no production code, visibility, or test
logic changes — only test relocation with adjusted `use` paths plus
removal of two now-unused inline helpers (`collection_page_count` and
`fetch_ignore_mirror`) that only the moved tests referenced.

Shared test helpers used across the new files (`open_test_db`,
`open_test_db_file`, `insert_collection`, `insert_page_with_raw_import`,
`page_id`, `quarantine_page`, `insert_knowledge_gap`, `insert_embedding_job`,
`collection_page_count`, `fetch_ignore_mirror`, `attach_collection`) are
lifted verbatim into `tests/common/collection_fixtures.rs`.

New files (each well under the 1500-LOC budget):
- tests/cli_collection_add.rs (5 tests) — add command refusals + attach
- tests/cli_collection_ignore.rs (5 tests) — ignore add/remove/clear/list
- tests/cli_collection_sync.rs (7 tests) — sync, audit, finalize-pending
- tests/cli_collection_reset.rs (3 tests) — restore-reset / reconcile-reset
- tests/cli_collection_platform_gates.rs (4 tests) — Windows refusals
- tests/cli_collection_routing.rs (2 tests) — put + list/info routing
- tests/common/collection_fixtures.rs (~190 LOC) — shared helpers

Inline residue (59 tests, 1810 LOC):
private items exercised include `add`, `remove`, `info`, `list`, `ignore`,
`audit`, `sync`, `restore`, `migrate_uuids`, `run_uuid_write_back`,
`quarantine_action`, `ignore_add/remove/clear/list`,
`ensure_unix_collection_command`, `describe_collection_status`,
`build_collection_info_output`, `collection_metrics`,
`read_initial_ignore_patterns`, `add_ignore_pattern_content`,
`remove_ignore_pattern_content`, `validate_cli_ignore_pattern`,
`validate_proposed_ignore_content`, `write_ignore_file_atomically`,
`cached_user_ignore_patterns`, `load_ignore_source`, `render_patterns_file`,
`ignore_content_contains_pattern`, `resolve_collection_root`,
`probe_root_writable`, `is_read_only_probe_error`, `load_collection_by_name`,
`writable_label`, `CollectionInfoOutput`, and `CollectionListRow`. Two
audit-surface tests also rely on `include_str!("collection.rs")` to grep
the production source against shape invariants.

Verification:
  baseline (parent f4b6c0a):  passed=1357 failed=0
  this commit:                passed=1357 failed=0

Both clippy channels green:
  cargo clippy --all-targets --locked -- -D warnings
  cargo clippy --all-targets --no-default-features \
    --features bundled,online-model --locked -- -D warnings

src/commands/collection.rs:  4279 LOC -> 3392 LOC (-887 LOC, -21%)

Part of OpenSpec change `extract-inline-tests-to-integration`.
