+++ src/commands/collection.rs
@@ -1563,10 +1563,25 @@ fn render_success(json: bool, value: serde_json::Value) -> Result<()> {
     Ok(())
 }
 
+// reason: white-box; needs `add`, `remove`, `info`, `list`, `ignore`,
+// `audit`, `sync`, `restore`, `migrate_uuids`, `run_uuid_write_back`,
+// `quarantine_action`, `ignore_add`, `ignore_remove`, `ignore_clear`,
+// `ignore_list`, `ensure_unix_collection_command`, `describe_collection_status`,
+// `build_collection_info_output`, `collection_metrics`,
+// `read_initial_ignore_patterns`, `add_ignore_pattern_content`,
+// `remove_ignore_pattern_content`, `validate_cli_ignore_pattern`,
+// `validate_proposed_ignore_content`, `write_ignore_file_atomically`,
+// `cached_user_ignore_patterns`, `load_ignore_source`, `render_patterns_file`,
+// `ignore_content_contains_pattern`, `resolve_collection_root`,
+// `probe_root_writable`, `is_read_only_probe_error`, `load_collection_by_name`,
+// `writable_label`, `CollectionInfoOutput`, and `CollectionListRow`. Two
+// audit-surface tests also use `include_str!("collection.rs")` to grep the
+// production source against shape invariants. Public-API tests for the
+// `run`/`CollectionAction::*` surfaces (add, ignore, sync/audit, reset,
+// platform gates, list/info/put routing) live under `tests/cli_collection_*.rs`.
 #[cfg(test)]
 mod tests {
     use super::*;
-    use crate::commands::put;
     use crate::core::{db, markdown};
     use sha2::{Digest, Sha256};
     use std::path::Path;
@@ -1716,29 +1731,6 @@ mod tests {
     }
 
     #[cfg(unix)]
-    fn collection_page_count(conn: &Connection, name: &str) -> i64 {
-        conn.query_row(
-            "SELECT COUNT(*)
-               FROM pages p
-               JOIN collections c ON c.id = p.collection_id
-              WHERE c.name = ?1 AND p.quarantined_at IS NULL",
-            [name],
-            |row| row.get(0),
-        )
-        .unwrap()
-    }
-
-    #[cfg(unix)]
-    fn fetch_ignore_mirror(conn: &Connection, name: &str) -> Option<String> {
-        conn.query_row(
-            "SELECT ignore_patterns FROM collections WHERE name = ?1",
-            [name],
-            |row| row.get(0),
-        )
-        .unwrap()
-    }
-
-    #[cfg(unix)]
     fn attach_collection(conn: &Connection, name: &str, root_path: &Path) {
         run(
             conn,
@@ -1757,220 +1749,12 @@ mod tests {
 
     #[cfg(unix)]
     #[test]
-    fn add_refuses_invalid_root_before_creating_collection_row() {
-        let conn = open_test_db();
-        let missing = PathBuf::from(r"D:\does-not-exist");
-
-        let error = run(
-            &conn,
-            CollectionAction::Add(CollectionAddArgs {
-                name: "work".to_owned(),
-                path: missing,
-                read_only: false,
-                writable: false,
-                write_quaid_id: false,
-                namespace: None,
-            }),
-            true,
-        )
-        .unwrap_err();
-
-        assert!(error.to_string().contains("collection root"));
-        let count: i64 = conn
-            .query_row(
-                "SELECT COUNT(*) FROM collections WHERE name = 'work'",
-                [],
-                |row| row.get(0),
-            )
-            .unwrap();
-        assert_eq!(count, 0);
-    }
-
-    #[test]
-    fn add_refuses_write_quaid_id_when_attach_is_read_only_before_creating_collection_row() {
-        let conn = open_test_db();
-        let root = tempfile::TempDir::new().unwrap();
-
-        let error = run(
-            &conn,
-            CollectionAction::Add(CollectionAddArgs {
-                name: "work".to_owned(),
-                path: root.path().to_path_buf(),
-                read_only: true,
-                writable: false,
-                write_quaid_id: true,
-                namespace: None,
-            }),
-            true,
-        )
-        .unwrap_err();
-
-        assert!(!error.to_string().is_empty());
-        let count: i64 = conn
-            .query_row(
-                "SELECT COUNT(*) FROM collections WHERE name = 'work'",
-                [],
-                |row| row.get(0),
-            )
-            .unwrap();
-        assert_eq!(count, 0);
-    }
-
-    #[cfg(unix)]
-    #[test]
-    fn add_refuses_invalid_ignore_before_creating_collection_row() {
-        let conn = open_test_db();
-        let root = tempfile::TempDir::new().unwrap();
-        fs::write(root.path().join(".quaidignore"), "[broken\n").unwrap();
-
-        let error = run(
-            &conn,
-            CollectionAction::Add(CollectionAddArgs {
-                name: "work".to_owned(),
-                path: root.path().to_path_buf(),
-                read_only: false,
-                writable: false,
-                write_quaid_id: false,
-                namespace: None,
-            }),
-            true,
-        )
-        .unwrap_err();
-
-        assert!(error.to_string().contains("invalid .quaidignore"));
-        assert!(error.to_string().contains("Fix .quaidignore and re-run"));
-        let count: i64 = conn
-            .query_row(
-                "SELECT COUNT(*) FROM collections WHERE name = 'work'",
-                [],
-                |row| row.get(0),
-            )
-            .unwrap();
-        assert_eq!(count, 0);
-    }
-
-    #[cfg(unix)]
-    #[test]
-    fn add_attaches_collection_and_cleans_short_lived_lease_residue() {
-        let (_dir, conn) = open_test_db_file();
-        let root = tempfile::TempDir::new().unwrap();
-        fs::write(
-            root.path().join("note.md"),
-            "---\ntitle: Note\ntype: note\n---\nhello\n",
-        )
-        .unwrap();
-
-        run(
-            &conn,
-            CollectionAction::Add(CollectionAddArgs {
-                name: "work".to_owned(),
-                path: root.path().to_path_buf(),
-                read_only: false,
-                writable: false,
-                write_quaid_id: false,
-                namespace: None,
-            }),
-            true,
-        )
-        .unwrap();
-
-        let row: (String, i64, Option<String>, i64, i64) = conn
-            .query_row(
-                "SELECT state, writable, active_lease_session_id,
-                        (SELECT COUNT(*) FROM collection_owners),
-                        (SELECT COUNT(*) FROM serve_sessions)
-                 FROM collections WHERE name = 'work'",
-                [],
-                |row| {
-                    Ok((
-                        row.get(0)?,
-                        row.get(1)?,
-                        row.get(2)?,
-                        row.get(3)?,
-                        row.get(4)?,
-                    ))
-                },
-            )
-            .unwrap();
-        assert_eq!(row.0, "active");
-        assert_eq!(row.1, 1);
-        assert!(row.2.is_none());
-        assert_eq!(row.3, 0);
-        assert_eq!(row.4, 0);
-
-        let page_count: i64 = conn
-            .query_row(
-                "SELECT COUNT(*) FROM pages p JOIN collections c ON c.id = p.collection_id WHERE c.name = 'work'",
-                [],
-                |row| row.get(0),
-            )
-            .unwrap();
-        assert_eq!(page_count, 1);
-    }
-
-    #[cfg(unix)]
-    #[test]
     fn capability_probe_leaves_no_residue_on_success() {
         let root = tempfile::TempDir::new().unwrap();
 
         let writable = probe_root_writable(root.path()).unwrap();
 
         assert!(writable);
-        let residue = fs::read_dir(root.path())
-            .unwrap()
-            .filter_map(|entry| entry.ok())
-            .any(|entry| {
-                entry
-                    .file_name()
-                    .to_string_lossy()
-                    .starts_with(".quaid-probe-")
-            });
-        assert!(!residue);
-    }
-
-    #[cfg(unix)]
-    #[test]
-    fn add_marks_collection_read_only_when_probe_hits_permission_denied() {
-        if rustix::process::geteuid().is_root() {
-            return;
-        }
-
-        let (_dir, conn) = open_test_db_file();
-        let root = tempfile::TempDir::new().unwrap();
-        fs::write(
-            root.path().join("note.md"),
-            "---\ntitle: Note\ntype: note\n---\nhello\n",
-        )
-        .unwrap();
-        let original_permissions = fs::metadata(root.path()).unwrap().permissions();
-        let mut read_only_permissions = original_permissions.clone();
-        read_only_permissions.set_mode(0o555);
-        fs::set_permissions(root.path(), read_only_permissions).unwrap();
-
-        let result = run(
-            &conn,
-            CollectionAction::Add(CollectionAddArgs {
-                name: "ro-vault".to_owned(),
-                path: root.path().to_path_buf(),
-                read_only: false,
-                writable: false,
-                write_quaid_id: false,
-                namespace: None,
-            }),
-            true,
-        );
-
-        fs::set_permissions(root.path(), original_permissions).unwrap();
-        result.unwrap();
-
-        let writable: i64 = conn
-            .query_row(
-                "SELECT writable FROM collections WHERE name = 'ro-vault'",
-                [],
-                |row| row.get(0),
-            )
-            .unwrap();
-        assert_eq!(writable, 0);
         let residue = fs::read_dir(root.path())
             .unwrap()
             .filter_map(|entry| entry.ok())
@@ -2063,58 +1847,6 @@ mod tests {
         };
         assert_eq!(info.writable_mode, "read-only");
         assert!(!info.writable);
-    }
-
-    #[test]
-    fn put_refuses_read_only_collection() {
-        let conn = open_test_db();
-        let root = tempfile::TempDir::new().unwrap();
-        let collection_id = insert_collection(&conn, "work", root.path());
-        conn.execute(
-            "UPDATE collections SET writable = 0 WHERE id = ?1",
-            [collection_id],
-        )
-        .unwrap();
-
-        let error = put::put_from_string(
-            &conn,
-            "work::notes/read-only",
-            "---\ntitle: Read Only\ntype: note\n---\nhello\n",
-            None,
-        )
-        .unwrap_err();
-
-        assert!(error.to_string().contains("CollectionReadOnlyError"));
-    }
-
-    #[test]
-    fn restore_reset_requires_confirm() {
-        let conn = db::open(":memory:").unwrap();
-        let error = run(
-            &conn,
-            CollectionAction::RestoreReset {
-                name: "work".to_owned(),
-                confirm: false,
-            },
-            true,
-        )
-        .unwrap_err();
-        assert!(error.to_string().contains("--confirm"));
-    }
-
-    #[test]
-    fn reconcile_reset_requires_confirm() {
-        let conn = db::open(":memory:").unwrap();
-        let error = run(
-            &conn,
-            CollectionAction::ReconcileReset {
-                name: "work".to_owned(),
-                confirm: false,
-            },
-            true,
-        )
-        .unwrap_err();
-        assert!(error.to_string().contains("--confirm"));
     }
 
     #[test]
@@ -2354,160 +2086,6 @@ mod tests {
         assert_eq!(total_raw_import_count(&conn, page_id), 1);
     }
 
-    #[cfg(unix)]
-    #[test]
-    fn ignore_add_updates_file_mirror_and_reconciles() {
-        let (_db_dir, conn) = open_test_db_file();
-        let root = tempfile::TempDir::new().unwrap();
-        fs::write(
-            root.path().join("note.md"),
-            "---\ntitle: Note\ntype: note\n---\nhello\n",
-        )
-        .unwrap();
-        attach_collection(&conn, "work", root.path());
-
-        run(
-            &conn,
-            CollectionAction::Ignore {
-                action: CollectionIgnoreAction::Add {
-                    name: "work".to_owned(),
-                    pattern: "note.md".to_owned(),
-                },
-            },
-            true,
-        )
-        .unwrap();
-
-        assert_eq!(
-            fs::read_to_string(root.path().join(".quaidignore")).unwrap(),
-            "note.md\n"
-        );
-        let mirror: Vec<String> =
-            serde_json::from_str(&fetch_ignore_mirror(&conn, "work").unwrap()).unwrap();
-        assert_eq!(mirror, vec!["note.md"]);
-        assert_eq!(collection_page_count(&conn, "work"), 0);
-    }
-
-    #[cfg(unix)]
-    #[test]
-    fn ignore_clear_removes_file_clears_mirror_and_reconciles() {
-        let (_db_dir, conn) = open_test_db_file();
-        let root = tempfile::TempDir::new().unwrap();
-        fs::write(
-            root.path().join("note.md"),
-            "---\ntitle: Note\ntype: note\n---\nhello\n",
-        )
-        .unwrap();
-        attach_collection(&conn, "work", root.path());
-        run(
-            &conn,
-            CollectionAction::Ignore {
-                action: CollectionIgnoreAction::Add {
-                    name: "work".to_owned(),
-                    pattern: "note.md".to_owned(),
-                },
-            },
-            true,
-        )
-        .unwrap();
-
-        run(
-            &conn,
-            CollectionAction::Ignore {
-                action: CollectionIgnoreAction::Clear {
-                    name: "work".to_owned(),
-                    confirm: true,
-                },
-            },
-            true,
-        )
-        .unwrap();
-
-        assert!(!root.path().join(".quaidignore").exists());
-        assert!(fetch_ignore_mirror(&conn, "work").is_none());
-        assert_eq!(collection_page_count(&conn, "work"), 1);
-    }
-
-    #[cfg(unix)]
-    #[test]
-    fn ignore_add_invalid_glob_refuses_without_disk_or_db_mutation() {
-        let conn = open_test_db();
-        let root = tempfile::TempDir::new().unwrap();
-        insert_collection(&conn, "work", root.path());
-
-        let error = run(
-            &conn,
-            CollectionAction::Ignore {
-                action: CollectionIgnoreAction::Add {
-                    name: "work".to_owned(),
-                    pattern: "[broken".to_owned(),
-                },
-            },
-            true,
-        )
-        .unwrap_err();
-
-        assert!(error.to_string().contains("Invalid glob pattern"));
-        assert!(!root.path().join(".quaidignore").exists());
-        assert!(fetch_ignore_mirror(&conn, "work").is_none());
-    }
-
-    #[cfg(unix)]
-    #[test]
-    fn ignore_remove_updates_file_and_mirror() {
-        let (_db_dir, conn) = open_test_db_file();
-        let root = tempfile::TempDir::new().unwrap();
-        fs::write(
-            root.path().join("note.md"),
-            "---\ntitle: Note\ntype: note\n---\nhello\n",
-        )
-        .unwrap();
-        attach_collection(&conn, "work", root.path());
-        run(
-            &conn,
-            CollectionAction::Ignore {
-                action: CollectionIgnoreAction::Add {
-                    name: "work".to_owned(),
-                    pattern: "note.md".to_owned(),
-                },
-            },
-            true,
-        )
-        .unwrap();
-        run(
-            &conn,
-            CollectionAction::Ignore {
-                action: CollectionIgnoreAction::Add {
-                    name: "work".to_owned(),
-                    pattern: "archive/**".to_owned(),
-                },
-            },
-            true,
-        )
-        .unwrap();
-
-        run(
-            &conn,
-            CollectionAction::Ignore {
-                action: CollectionIgnoreAction::Remove {
-                    name: "work".to_owned(),
-                    pattern: "note.md".to_owned(),
-                },
-            },
-            true,
-        )
-        .unwrap();
-
-        assert_eq!(
-            fs::read_to_string(root.path().join(".quaidignore")).unwrap(),
-            "archive/**\n"
-        );
-        let mirror: Vec<String> =
-            serde_json::from_str(&fetch_ignore_mirror(&conn, "work").unwrap()).unwrap();
-        assert_eq!(mirror, vec!["archive/**"]);
-        assert_eq!(collection_page_count(&conn, "work"), 1);
-    }
-
     #[test]
     fn ignore_mutations_refuse_while_collection_is_restoring() {
         let conn = open_test_db();
@@ -2625,410 +2203,6 @@ mod tests {
             helper.contains("\"raw_imports_deleted\": raw_imports_deleted"),
             "audit helper must report the deleted-row count in its rendered output"
         );
-    }
-
-    #[cfg(not(unix))]
-    #[test]
-    fn add_refuses_windows_platform() {
-        let conn = open_test_db();
-        let root = tempfile::TempDir::new().unwrap();
-
-        let error = run(
-            &conn,
-            CollectionAction::Add(CollectionAddArgs {
-                name: "work".to_owned(),
-                path: root.path().to_path_buf(),
-                read_only: false,
-                writable: false,
-                write_quaid_id: false,
-                namespace: None,
-            }),
-            true,
-        )
-        .unwrap_err();
-
-        assert!(error.to_string().contains("UnsupportedPlatformError"));
-    }
-
-    #[cfg(not(unix))]
-    #[test]
-    fn sync_refuses_windows_platform() {
-        let conn = open_test_db();
-
-        let error = run(
-            &conn,
-            CollectionAction::Sync(CollectionSyncArgs {
-                name: "work".to_owned(),
-                remap_root: None,
-                finalize_pending: false,
-                online: false,
-            }),
-            true,
-        )
-        .unwrap_err();
-
-        assert!(error.to_string().contains("UnsupportedPlatformError"));
-    }
-
-    #[cfg(not(unix))]
-    #[test]
-    fn restore_refuses_windows_platform() {
-        let conn = open_test_db();
-        let target = tempfile::TempDir::new().unwrap();
-
-        let error = run(
-            &conn,
-            CollectionAction::Restore(CollectionRestoreArgs {
-                name: "work".to_owned(),
-                target: target.path().to_path_buf(),
-                online: false,
-            }),
-            true,
-        )
-        .unwrap_err();
-
-        assert!(error.to_string().contains("UnsupportedPlatformError"));
-    }
-
-    #[cfg(not(unix))]
-    #[test]
-    fn audit_refuses_windows_platform() {
-        let conn = open_test_db();
-
-        let error = run(
-            &conn,
-            CollectionAction::Audit(CollectionAuditArgs {
-                name: "work".to_owned(),
-                raw_imports_gc: false,
-            }),
-            true,
-        )
-        .unwrap_err();
-
-        assert!(error.to_string().contains("UnsupportedPlatformError"));
-    }
-
-    #[cfg(unix)]
-    #[test]
-    fn sync_finalize_pending_uses_external_finalize_path() {
-        let (_db_dir, conn) = open_test_db_file();
-        let temp = tempfile::TempDir::new().unwrap();
-        let pending_root = temp.path().join("restored");
-        fs::create_dir_all(&pending_root).unwrap();
-        let collection_id = insert_collection(&conn, "work", temp.path());
-        conn.execute(
-            "UPDATE collections
-             SET state = 'restoring',
-                 pending_root_path = ?2,
-                 pending_restore_manifest = '{\"entries\":[]}',
-                 restore_command_id = 'restore-1',
-                 pending_command_heartbeat_at = datetime('now', '-120 seconds')
-             WHERE id = ?1",
-            rusqlite::params![collection_id, pending_root.display().to_string()],
-        )
-        .unwrap();
-
-        run(
-            &conn,
-            CollectionAction::Sync(CollectionSyncArgs {
-                name: "work".to_owned(),
-                remap_root: None,
-                finalize_pending: true,
-                online: false,
-            }),
-            true,
-        )
-        .unwrap();
-
-        let row: (String, String, i64, Option<String>) = conn
-            .query_row(
-                "SELECT state, root_path, needs_full_sync, pending_root_path
-                  FROM collections WHERE id = ?1",
-                [collection_id],
-                |row| Ok((row.get(0)?, row.get(1)?, row.get(2)?, row.get(3)?)),
-            )
-            .unwrap();
-        assert_eq!(row.0, "active");
-        assert_eq!(row.1, pending_root.display().to_string());
-        assert_eq!(row.2, 0);
-        assert!(row.3.is_none());
-    }
-
-    #[cfg(unix)]
-    #[test]
-    fn audit_runs_full_hash_reconcile_and_optional_raw_import_gc() {
-        let (_db_dir, conn) = open_test_db_file();
-        let root = tempfile::TempDir::new().unwrap();
-        let collection_id = insert_collection(&conn, "work", root.path());
-        let relative_path = "notes/a.md";
-        let file_path = root.path().join(relative_path);
-        fs::create_dir_all(file_path.parent().unwrap()).unwrap();
-        let uuid = Uuid::now_v7().to_string();
-        let raw_bytes =
-            format!("---\nmemory_id: {uuid}\nslug: notes/a\ntitle: A\ntype: concept\n---\nBody.\n");
-        fs::write(&file_path, raw_bytes.as_bytes()).unwrap();
-        let page_id = insert_page_with_raw_import(
-            &conn,
-            collection_id,
-            "notes/a",
-            &uuid,
-            raw_bytes.as_bytes(),
-            relative_path,
-        );
-        conn.execute(
-            "UPDATE file_state
-             SET last_full_hash_at = datetime('now', '-8 days')
-             WHERE collection_id = ?1",
-            [collection_id],
-        )
-        .unwrap();
-        conn.execute(
-            "INSERT INTO raw_imports (page_id, import_id, is_active, raw_bytes, file_path, created_at)
-             VALUES (?1, ?2, 0, ?3, ?4, '2000-01-01T00:00:00Z')",
-            rusqlite::params![page_id, Uuid::now_v7().to_string(), b"old", relative_path],
-        )
-        .unwrap();
-
-        run(
-            &conn,
-            CollectionAction::Audit(CollectionAuditArgs {
-                name: "work".to_owned(),
-                raw_imports_gc: true,
-            }),
-            true,
-        )
-        .unwrap();
-
-        let row: (Option<String>, i64) = conn
-            .query_row(
-                "SELECT last_sync_at,
-                        (SELECT COUNT(*)
-                         FROM raw_imports
-                         WHERE page_id = ?2 AND is_active = 0)
-                 FROM collections
-                 WHERE id = ?1",
-                rusqlite::params![collection_id, page_id],
-                |row| Ok((row.get(0)?, row.get(1)?)),
-            )
-            .unwrap();
-        assert!(row.0.is_some());
-        assert_eq!(row.1, 0);
-    }
-
-    #[cfg(unix)]
-    #[test]
-    fn sync_without_flags_requires_active_root_collection() {
-        let conn = db::open(":memory:").unwrap();
-        let temp = tempfile::TempDir::new().unwrap();
-        let collection_id = insert_collection(&conn, "work", temp.path());
-        conn.execute(
-            "UPDATE collections SET state = 'detached' WHERE id = ?1",
-            [collection_id],
-        )
-        .unwrap();
-        let error = run(
-            &conn,
-            CollectionAction::Sync(CollectionSyncArgs {
-                name: "work".to_owned(),
-                remap_root: None,
-                finalize_pending: false,
-                online: false,
-            }),
-            true,
-        )
-        .unwrap_err();
-
-        assert!(error
-            .to_string()
-            .contains("PlainSyncActiveRootRequiredError"));
-    }
-
-    #[cfg(unix)]
-    #[test]
-    fn sync_without_flags_refuses_restore_in_progress_state() {
-        let conn = db::open(":memory:").unwrap();
-        let temp = tempfile::TempDir::new().unwrap();
-        let collection_id = insert_collection(&conn, "work", temp.path());
-        conn.execute(
-            "UPDATE collections SET state = 'restoring' WHERE id = ?1",
-            [collection_id],
-        )
-        .unwrap();
-
-        let error = run(
-            &conn,
-            CollectionAction::Sync(CollectionSyncArgs {
-                name: "work".to_owned(),
-                remap_root: None,
-                finalize_pending: false,
-                online: false,
-            }),
-            true,
-        )
-        .unwrap_err();
-
-        let state: String = conn
-            .query_row(
-                "SELECT state FROM collections WHERE id = ?1",
-                [collection_id],
-                |row| row.get(0),
-            )
-            .unwrap();
-        assert!(error.to_string().contains("RestoreInProgressError"));
-        assert_eq!(state, "restoring");
-    }
-
-    #[cfg(unix)]
-    #[test]
-    fn sync_without_flags_does_not_finalize_pending_restore_state() {
-        let conn = db::open(":memory:").unwrap();
-        let temp = tempfile::TempDir::new().unwrap();
-        let pending_root = temp.path().join("restored");
-        fs::create_dir_all(&pending_root).unwrap();
-        let collection_id = insert_collection(&conn, "work", temp.path());
-        conn.execute(
-            "UPDATE collections
-             SET state = 'restoring',
-                 pending_root_path = ?2,
-                 pending_restore_manifest = '{\"entries\":[]}',
-                 restore_command_id = 'restore-1',
-                 pending_command_heartbeat_at = strftime('%Y-%m-%dT%H:%M:%SZ', 'now')
-             WHERE id = ?1",
-            rusqlite::params![collection_id, pending_root.display().to_string()],
-        )
-        .unwrap();
-
-        let error = run(
-            &conn,
-            CollectionAction::Sync(CollectionSyncArgs {
-                name: "work".to_owned(),
-                remap_root: None,
-                finalize_pending: false,
-                online: false,
-            }),
-            true,
-        )
-        .unwrap_err();
-
-        let row: (String, Option<String>, Option<String>) = conn
-            .query_row(
-                "SELECT state, pending_root_path, restore_command_id
-                 FROM collections WHERE id = ?1",
-                [collection_id],
-                |row| Ok((row.get(0)?, row.get(1)?, row.get(2)?)),
-            )
-            .unwrap();
-        assert!(error.to_string().contains("RestorePendingFinalizeError"));
-        assert_eq!(row.0, "restoring");
-        assert_eq!(row.1.as_deref(), Some(pending_root.to_str().unwrap()));
-        assert_eq!(row.2.as_deref(), Some("restore-1"));
-    }
-
-    #[cfg(unix)]
-    #[test]
-    fn sync_without_flags_refuses_restore_integrity_blocked_state() {
-        let conn = db::open(":memory:").unwrap();
-        let temp = tempfile::TempDir::new().unwrap();
-        let collection_id = insert_collection(&conn, "work", temp.path());
-        conn.execute(
-            "UPDATE collections
-             SET state = 'restoring',
-                 integrity_failed_at = '2026-04-23T00:00:00Z'
-             WHERE id = ?1",
-            [collection_id],
-        )
-        .unwrap();
-
-        let error = run(
-            &conn,
-            CollectionAction::Sync(CollectionSyncArgs {
-                name: "work".to_owned(),
-                remap_root: None,
-                finalize_pending: false,
-                online: false,
-            }),
-            true,
-        )
-        .unwrap_err();
-
-        let row: (String, Option<String>) = conn
-            .query_row(
-                "SELECT state, integrity_failed_at FROM collections WHERE id = ?1",
-                [collection_id],
-                |row| Ok((row.get(0)?, row.get(1)?)),
-            )
-            .unwrap();
-        assert!(error.to_string().contains("RestoreIntegrityBlockedError"));
-        assert_eq!(row.0, "restoring");
-        assert_eq!(row.1.as_deref(), Some("2026-04-23T00:00:00Z"));
-    }
-
-    #[cfg(unix)]
-    #[test]
-    fn sync_without_flags_does_not_clear_integrity_or_reconcile_halt_markers() {
-        let conn = db::open(":memory:").unwrap();
-        let temp = tempfile::TempDir::new().unwrap();
-        let collection_id = insert_collection(&conn, "work", temp.path());
-        conn.execute(
-            "UPDATE collections
-             SET integrity_failed_at = '2026-04-23T00:00:00Z',
-                 reconcile_halted_at = '2026-04-23T00:05:00Z',
-                 reconcile_halt_reason = 'duplicate_uuid'
-             WHERE id = ?1",
-            [collection_id],
-        )
-        .unwrap();
-
-        let error = run(
-            &conn,
-            CollectionAction::Sync(CollectionSyncArgs {
-                name: "work".to_owned(),
-                remap_root: None,
-                finalize_pending: false,
-                online: false,
-            }),
-            true,
-        )
-        .unwrap_err();
-
-        let row: (Option<String>, Option<String>, Option<String>) = conn
-            .query_row(
-                "SELECT integrity_failed_at, reconcile_halted_at, reconcile_halt_reason
-                 FROM collections WHERE id = ?1",
-                [collection_id],
-                |row| Ok((row.get(0)?, row.get(1)?, row.get(2)?)),
-            )
-            .unwrap();
-        assert!(error.to_string().contains("ReconcileHaltedError"));
-        assert_eq!(row.0.as_deref(), Some("2026-04-23T00:00:00Z"));
-        assert_eq!(row.1.as_deref(), Some("2026-04-23T00:05:00Z"));
-        assert_eq!(row.2.as_deref(), Some("duplicate_uuid"));
-    }
-
-    #[test]
-    fn run_routes_ignore_list_action() {
-        let conn = open_test_db();
-        let root = tempfile::TempDir::new().unwrap();
-        let collection_id = insert_collection(&conn, "work", root.path());
-        conn.execute(
-            "UPDATE collections SET ignore_patterns = ?2 WHERE id = ?1",
-            rusqlite::params![
-                collection_id,
-                serde_json::to_string(&vec!["private/**"]).unwrap()
-            ],
-        )
-        .unwrap();
-
-        run(
-            &conn,
-            CollectionAction::Ignore {
-                action: CollectionIgnoreAction::List {
-                    name: "work".to_owned(),
-                },
-            },
-            true,
-        )
-        .unwrap();
     }
 
     #[test]
@@ -3607,56 +2781,6 @@ mod tests {
     }
 
     #[test]
-    fn restore_reset_and_reconcile_reset_succeed_when_confirmed() {
-        let conn = open_test_db();
-        let root = tempfile::TempDir::new().unwrap();
-        let collection_id = insert_collection(&conn, "work", root.path());
-        conn.execute(
-            "UPDATE collections
-             SET state = 'restoring',
-                 pending_root_path = 'D:\\vault\\restored',
-                 integrity_failed_at = '2026-04-23T00:00:00Z',
-                 reconcile_halted_at = '2026-04-23T00:05:00Z',
-                 reconcile_halt_reason = 'duplicate_uuid'
-             WHERE id = ?1",
-            [collection_id],
-        )
-        .unwrap();
-
-        run(
-            &conn,
-            CollectionAction::RestoreReset {
-                name: "work".to_owned(),
-                confirm: true,
-            },
-            true,
-        )
-        .unwrap();
-        run(
-            &conn,
-            CollectionAction::ReconcileReset {
-                name: "work".to_owned(),
-                confirm: true,
-            },
-            true,
-        )
-        .unwrap();
-
-        let row: (String, Option<String>, Option<String>, Option<String>) = conn
-            .query_row(
-                "SELECT state, pending_root_path, integrity_failed_at, reconcile_halted_at
-                 FROM collections WHERE id = ?1",
-                [collection_id],
-                |row| Ok((row.get(0)?, row.get(1)?, row.get(2)?, row.get(3)?)),
-            )
-            .unwrap();
-        assert_eq!(row.0, "active");
-        assert!(row.1.is_none());
-        assert!(row.2.is_none());
-        assert!(row.3.is_none());
-    }
-
-    #[test]
     fn list_and_info_plain_text_helpers_return_ok_for_seeded_collection() {
         let conn = open_test_db();
         let root = tempfile::TempDir::new().unwrap();
@@ -3675,41 +2799,6 @@ mod tests {
 
         list(&conn, false).unwrap();
         info(&conn, "work", false).unwrap();
-    }
-
-    #[test]
-    fn run_routes_list_and_info_json_helpers_for_seeded_collection() {
-        let conn = open_test_db();
-        let root = tempfile::TempDir::new().unwrap();
-        let collection_id = insert_collection(&conn, "work", root.path());
-        let page_id = insert_page_with_raw_import(
-            &conn,
-            collection_id,
-            "notes/info",
-            &Uuid::now_v7().to_string(),
-            b"---\ntitle: Info\ntype: note\n---\nbody\n",
-            "notes/info.md",
-        );
-        quarantine_page(&conn, page_id, "2026-04-28T00:00:00Z");
-        insert_embedding_job(&conn, page_id, "failed", 2);
-        conn.execute(
-            "UPDATE collections
-             SET integrity_failed_at = '2026-04-28T00:00:00Z',
-                 ignore_parse_errors = 'line 1 raw=\"[broken\" error=Invalid glob pattern'
-             WHERE id = ?1",
-            [collection_id],
-        )
-        .unwrap();
-
-        run(&conn, CollectionAction::List, true).unwrap();
-        run(
-            &conn,
-            CollectionAction::Info {
-                name: "work".to_owned(),
-            },
-            true,
-        )
-        .unwrap();
     }
 
     #[test]
